+++ load.py
@@ -157,8 +157,8 @@ def plugin_app(parent):
         title2 = tk.Label(this.frame, text="New version available", fg="blue", cursor="hand2")
         title2.grid(row=0, column=1, sticky=tk.W, )
         title2.bind("<Button-1>", lambda e: webbrowser.open_new("https://github.com/aussig/BGS-Tally/releases"))
-    tk.Button(this.frame, text='Latest Tick Data', command=display_todaydata).grid(row=1, column=0, padx=3)
-    tk.Button(this.frame, text='Earlier Tick Data', command=display_yesterdaydata).grid(row=1, column=1, padx=3)
+    tk.Button(this.frame, text='Latest Tally', command=display_todaydata).grid(row=1, column=0, padx=3)
+    tk.Button(this.frame, text='Previous Tally', command=display_yesterdaydata).grid(row=1, column=1, padx=3)
     tk.Label(this.frame, text="Status:").grid(row=2, column=0, sticky=tk.W)
     tk.Label(this.frame, text="Last Tick:").grid(row=3, column=0, sticky=tk.W)
     this.StatusLabel = tk.Label(this.frame, text=this.Status.get())
@@ -425,11 +425,11 @@ def generate_discord_text(data):
     
 
 def display_todaydata():
-    display_data("Latest Tick Data", this.TodayData)
+    display_data("Latest Tally", this.TodayData)
 
 
 def display_yesterdaydata():
-    display_data("Earlier Tick Data", this.YesterdayData)
+    display_data("Previous Tally", this.YesterdayData)
 
 
 def copy_to_clipboard(form, discord_text):