+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changes:
 
 * Now include a lot more non-violent mission types when counting missions for a faction in the `Election` state (gathering a full list of non-violent mission types is still a work in progress). 
+* Rename buttons and windows to 'Latest Tally' and 'Previous Tally'.
 
 
 ## v1.1.1 - 2021-08-31